requirements.txt
# LangChain dependencies
langchain==1.0.7
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.5
langchain-openai==1.0.3
langchain-text-splitters==1.0.0

# Vector Database (Weaviate)
weaviate-client>=4.0.0

# Knowledge Graph (Neo4j)
neo4j>=5.0.0

# LVLM Integration (Ollama Python client)
ollama>=0.1.0

# Image Processing
Pillow>=10.0.0

# Evaluation Framework
deepeval>=0.21.0

# Note: Ollama application should be installed separately on your system
# See: https://ollama.ai/
# After installing Ollama, pull a LLaVA model:
#   ollama pull llava-phi3
